+++ plugins/subtitles/subtitle-utils.js
@@ -6,14 +6,14 @@ export function youTubeTranscript2SRT(transcript) {
 
   transcript.forEach((t, index, array) => {
 
-    const text = t.t._;
-    const startTime = Number(t.t['$'].start).toFixed();
-    const duration = Number(t.t.$.dur).toFixed();
+    const text = t._;
+    const startTime = Number(t.$.start).toFixed();
+    const duration = Number(t.$.dur).toFixed();
     const endTime = Number(startTime) + Number(duration);
     const startTimeInHHMMSS = secondsToHHMMSS(startTime);
     let endTimeInHHMMSS = secondsToHHMMSS(endTime);
     const isNotLastElement = index < array.length - 1;
-    const nextStartTime = isNotLastElement ? Number(array[index + 1].t['$'].start).toFixed() : Infinity;
+    const nextStartTime = isNotLastElement ? Number(array[index + 1].$.start).toFixed() : Infinity;
     const srtIndex = index + 1;
 
     // FIX TIMING