+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `extract_ner` pattern for LLM
 - Add spaCy 'Named Entity Recognition' plugin: `atlas -u ner`
 - Add General-purpose `atlas -u converter` plugin with support for HH:MM:SS-to-Seconds and vice versa
+- Add `--format=srt` option to YouTube transcript downloader (`--youtube`)
 
 ## [0.1.7] - 11/09/2026